LayoutTests/http/tests/webgpu/webgpu/api/operation/compute_pipeline/entry_point_name.html


                <script>
                    if (window.testRunner)
                        testRunner.waitUntilDone();
                    window.location.href = "/webgpu/cts.https.html?q=webgpu:api,operation,compute_pipeline,entry_point_name:*";
                </script>
            
